+++ client/app/center/center.controller.js
@@ -10,7 +10,7 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
                                                  Sticker, jndPubSub, jndKeyCode, DeskTopNotificationBanner,
                                                  MessageCollection, MessageSendingCollection, AnalyticsHelper,
                                                  Announcement, TopicMessageCache, NotificationManager, Dialog, RendererUtil,
-                                                 JndUtil, HybridAppHelper) {
+                                                 JndUtil, HybridAppHelper, TopicInvitedFlagMap, EntityMapManager) {
 
   //console.info('::[enter] centerpanelController', $state.params.entityId);
   var _scrollHeightBefore;
@@ -47,7 +47,9 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
   var deferredObject = {
     getMessage: null,
     postMessage: null,
-    updateMessage: null
+    updateMessage: null,
+    updateMessageMarker: null,
+    getRoomInformation: null
   };
 
   var firstMessageId;             // 현재 엔티티(토픽, DM)의 가장 위 메세지 아이디.
@@ -141,10 +143,11 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
   function _init() {
     centerService.preventChatWithMyself(entityId);
     $rootScope.isIE9 = centerService.isIE9();
+    TopicInvitedFlagMap.remove(entityId);
 
     $scope.$on('$destroy', _onDestroy);
     $scope.$on('$viewContentLoaded', _onViewContentLoaded);
-    
+
     //entity 리스트 load 가 완료되지 않았다면 dataInitDone 이벤트를 기다린다
     if (publicService.isInitDone()) {
       _initializeListeners();
@@ -1028,7 +1031,8 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
   //  Updating message marker for current entity.
   function updateMessageMarker() {
     //console.log('::updateMessageMarker', lastMessageId);
-    messageAPIservice.updateMessageMarker(entityId, entityType, lastMessageId)
+    deferredObject.updateMessageMarker = $q.defer();
+    messageAPIservice.updateMessageMarker(entityId, entityType, lastMessageId, deferredObject.updateMessageMarker)
       .success(function(response) {
         memberService.setLastReadMessageMarker(_getEntityId(), lastMessageId);
         //log('----------- successfully updated message marker for entity name ' + $scope.currentEntity.name + ' to ' + lastMessageId);
@@ -1227,11 +1231,12 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
   }
 
   function setCommentFocus(file) {
+    var writer;
     if ($state.params.itemId != file.id) {
       $rootScope.setFileDetailCommentFocus = true;
-
+      writer = EntityMapManager.get('member', file.writerId);
       $state.go('files', {
-        userName    : file.writer.name,
+        userName    : writer.name,
         itemId      : file.id
       });
     } else {
@@ -1569,8 +1574,8 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
    */
   function _getCurrentRoomInfo() {
     var currentRoomId = _getEntityId();
-
-    messageAPIservice.getRoomInformation(currentRoomId)
+    deferredObject.getRoomInformation = $q.defer();
+    messageAPIservice.getRoomInformation(currentRoomId, deferredObject.getRoomInformation)
       .success(function(response) {
         _initMarkers(response.markers);
         hasRetryGetRoomInfo = false;
@@ -1656,7 +1661,7 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
    */
   function _onChangeSticker(event, item) {
     _sticker = item;
-    _setChatInputFocus();
+    setTimeout(_setChatInputFocus);
   }
 
   /**
@@ -1688,9 +1693,9 @@ app.controller('centerpanelController', function($scope, $rootScope, $state, $fi
     var messageId;
     var linkPreview;
     var timeoutCaller;
-
+    deferredObject.getMessage = $q.defer();
     messageAPIservice
-      .getMessage(memberService.getTeamId(), data.message.id)
+      .getMessage(memberService.getTeamId(), data.message.id, {}, deferredObject.getMessage)
       .success(function(response) {
         messageId = response.id;
         linkPreview = response.linkPreview;